Update button display logic
When you would click on digits and operators, it would just add to the
display. Now when you click on digits, it gets added to the left operand
span. Then you choose an operator which gets put in the operator span.
Then if you click on digits again, they will get put in the right
operand span.