Refactor Code
Update several functions and delete others. I first added data
attribute to each cell, then updated the makeMove() function to get
that data attribute and use it to push a variable in gameboard array.
This now has nothing to do with setting the inner text content. It now
pushes a variable into array, then loops through that array and displays
it